refactor: create `Thumbnail` component from index `Screenshot`
This lets us reuse it easily in other pages.

In the process, the inline `<script>` content was mostly pulled into a
separate function, which is then injected into the `<script>` via
`Function.prototype.toString`. This means the code for the `<script>`
benefits from linting, type-checking, formatting, etc.

This mostly works with the exception of a weird `__name(...)` function
that shows up in the result. To work around that, a no-op `__name(...)`
function is also defined in the `<script>` tag.